refactor: add account service class (#3612)
* chore: add account service


* chore: fix comment


* chore: edit comments


* chore: resolve comments


* chore: resolve comments


* chore: resolve comments


---------
